Make currentUseIndex a required prop again
Considered choosing a default behaviour for the back button, e.g. if there is no currentUseIndex, back button takes user to lowest priority use.  But decided that this could have unintended consequences / confusing behaviour at edge cases.

Decided to make the currentUseIndex prop required so there is no default behaviour if the parameter is not set.  Other pages that link to this page must decide what the Back button action is.

Choosing a default behaviour remains an option in the future should not having default behaviour become an issue.