Fixes Non-Descriptive Links in Context (2)
Updated `AnchorLink` and `WarningLink` components to optionally receive
a description prop which, if present, will render a visually hidden span
that aids people using screen readers to understand the purpose of the
link without context. Used this change to add descriptions to the change
and delete links in page `register-a-beacon/additional-beacon-use`.