Permit access to OpenSearch via the public internet
It does not appear possible to create an ALB forwarding rule from the
public Internet to an OpenSearch domain.  (This is the current way that
access via the public Internet is permitted to the Webapp and Service)

This commit changes the ingress rule of the security group that contains
OpenSearch to permit unrestricted ingress from the public Internet.

This is intended to be a temporary solution to allow us to confirm
that test data has been correctly seeded to OpenSearch.

Prior to pushing further in the pipeline, OpenSearch should be
 configured with access control.