Remove plugin that provides git info on actuator endpoint (2)
This plugin was installed to provide the caller with the SHA of the commit
that deployed the Service.

If we move to versioned deployments, this is no longer necessary. We'll
know which version of the Service is deployed by looking at ECS, not be querying the application.

Remove associated test.