Move CodeQL check to a pre-req for release to Production only
The CodeQL action acts on a _repo_.  The only way to create discrete CodeQL
checks for each subdirectory is to have pattern matching excluding other
directories and running CodeQL multiple times.

This commit removes CodeQL responsibilities from the publication of images
and instead makes it a pre-req of pushing to production only.

This will speed up images getting to staging, so testing can be performed in
the cloud concurrently to CodeQL checks.