When a customer updates a Beacon then reindex in Opensearch.
Instead of partially updating the document we will replace the document.
We have decided to do this because the update transaction crosses
multiple aggregate boundaries. If, in the future, we decide to update
each aggregate individually then this can be adjusted to be a partial
update to the document for each aggregate.