Roll back to v1.0
Putting the main branch in the same state as was deployed in v1.0.

I needed to do this to set .tfvars that give staging infrastructure parity with production for a pen test.

I did not want to risk performing a 'terraform apply' operation to the staging environment from my laptop
and bring down the environment during a pentest.

Instead I chose to run the infra changes through the CI/CD pipeline, which required reverting code to the
last deployed state.

See https://stackoverflow.com/a/33890073/11948408 for git commands run.