Add auth headers on /search/ location.
Adding the auth headers for  application_user with readall permissions
allows the backoffice application to query OpenSearch via the proxy
without having to provide the basic auth headers itself. All requests
from the load balancer to the proxy require the client to be
authenticated via Azure AD so this is safe.