Set env var to satisfy startup script in local dev
Because the env var is evaluated in the nginx config template, we have
no way of cleanly telling nginx not to evaluate the env var per
environment.  No setting of defaults, for example.

We decided the least bad option was to pass in a blank env var in the
local docker compose.