Set ALB to deny unauthenticated req's to OpenSearch Proxy
This was previously set to the default of "authenticate", thus
redirected requests to Microsoft's servers for the user to provide
credentials.

In our use case, this is counter-productive as this listener rule is
used by _applications_, not users.  Setting to 'deny' prevents
confusing error messages about CORS, redirects etc. when the request
is from a browser session that is not authenticated.

Co-authored-by: Olly Swanson <olly.swanson95@gmail.com>