Define a hook to retrieve authentication config and state
ClientId and TenantId need to be set at runtime because different
environments use a different Azure AD tenant.

This commit creates a hook to get the Azure AD tenant and client IDs
from an endpoint served by NGINX, allowing this configuration to be
made at runtime (as opposed to build time).

Co-authored-by: Olly Swanson <olly.swanson95@gmail.com>