Add health checks to service discovery
We think a bug discovered on dev/staging was due to incomplete
de-registering of the Service with AWS CloudMap, possibly when
Terraform's state lock became corrupted.

This commit adds health checks to all service discovery services, which
we think will cause unhealthy services to deregister and/or all traffic
to only be sent to healthy services.

Co-authored-by: Olly Swanson <olly.swanson95@gmail.com>