Remove wait_for_steady_state from services in the application cluster
We found that wait_for_steady_state is sometimes exceeded by long
startup times and the 5 minute rule for deregistering with the load
balancer.  (This is a limitation of the AWS API, see
https://docs.aws.amazon.com/cli/latest/reference/ecs/wait/services-stable.html)

This commit removes the wait_for_steady_state flag as this need is now
catered for using wait-for-ecs-deployment.sh in GitHub Actions.

wait_for_steady_state is left on in opensearch_proxy due to the very
low startup time for that container (it's just a proxy).

Co-authored-by: Olly Swanson <olly.swanson95@gmail.com>