Reference SSL certificate that includes wildcard for subdomain.
The previously registered certificate secures the domains
dev/staging/production.406beacons.com, with the certificate that secures
register-406-beacons.service.gov.uk registered manually.

This will remove the need for the manual registration and remove the
certificate for production.406beacons.com, reducing the certificates
registered with the load balancer from 2 to 1.

Co-authored-by: Zack Adlington <zadlington@gmail.com>