Remove MirageJS as a dependency, remove NotesIntegration test.
We decided to drop stubbing with Mirage in favour of running the local
setup with the supporting containers.

The Notes integration test relied on the mock server provided by Mirage
but was malfunctioning, so has been removed in favour of reliance on
unit testing. In future this use case should be covered by automated
acceptance testing against the cloud environment.

Co-authored-by: Zack Adlington <54271433+zackads@users.noreply.github.com>