Reduced deregistration delay for alb target groups from 300s to 60s.
Due to "false failures" of the CI/CD pipeline caused by the wait for
steady state flag, we have reduced the time that it takes for each
service to deregister from the alb target groups from 300s to 60s.

This should reduce the time to taken for each service to achieve steady
state.

Discussed in: 6c971af4 (Replace wait_for_steady_state flag in application services, 2022-03-14)
Co-authored-by: Zack Adlington <54271433+zackads@users.noreply.github.com>