Update deploy workflow to "check out" a specific ref.
The current deploy workflow checks out the HEAD of main. This means that
any artifacts that are deployed won't necessarily be deployed with the
version of the infrastructure that they are expecting.

This means that "deploy dev" will continue to use HEAD while "deploy
staging/prod" will use the same tag to deploy infrastructure and
artifacts.

Co-authored-by: Zack Adlington <54271433+zackads@users.noreply.github.com>