Remove @SpringBatchTest annotation and replace with Configuration.
@SpringBatchTest creates Test util beans and autowires their
dependencies without the ability to qualify them. We therefore remove
the annotation and provide the test utils in the application context
ourselves.

We had to make this change as the addition of a second job created
ambiguity that @SpringBatchTest was unable to resolve.

Co-authored by: Zack Adlington <54271433+zackads@users.noreply.github.com>