Assert on content type of spreadsheet response
Expect that the /export/excel endpoint will respond with a binary.

We chose not to assert on the structure of the returned spreadsheet
because (a) the shape/structure is tested in the batch job and (b) we
want to focus on testing the download works for Backoffice users.  We
don't feel the time spent deconstructing the spreadsheet in the Spring
integration test will be well spent.

Co-authored-by: Olly Swanson <42551149+ollyswanson@users.noreply.github.com>