Fix mock flakiness in unit test
We discovered Mockito @InjectMocks was not reliably injecting the right
mocks into the right properties.  The presence of two JobLaunchers
in the ExportService (one async, one sync) was causing mocks to be
injected in different orders, causing tests to fail sometimes depending
on the ordering of injection.

This commit replaces autowiring of mocks with explicit instantiation and
injection with code.

Co-authored-by: Olly Swanson <42551149+ollyswanson@users.noreply.github.com>