Remove subdirectory mount point from Service volume
Was causing a 'No such file or directory' error.   The EFS filesystem
the volume references is for the exclusive use of the Service, so
removing this parameter sets the mount point to root '/'.

Going forward, we'll need to keep the directory structure of the
Service's filesystem a concern of the Service.