Add extra fields to `SpreadsheetRow` (from BeaconOwner and BeaconUse)
In order to keep the row flat we have decided to handle the many-to-one
relationship of Beacon -> BeaconUse by concatenating the intersecting
fields of each BeaconUse into a single string. This may not be the
ideal approach, but because the "many" side of the relationship is
unconstrained, this is the simplest approach.