Simplify Dockerfile
node:16-alpine was having issues running the postinstall script in the
build

This commit copies the webapp/ directory into the base image rather than
cherry picking files here and there.

The base image is also the non-alpine version to reduce chances of
incompatibility in future.  There is no advantage to using an alpine
image for builds, as that image is just thrown away.