Prevent Dependabot from submitting Webapp PRs
We've identified that the Webapp is not fit for purpose: it has an
unacceptably high error rate and we have recommended to the MCA that its
functionality is migrated into Spring Boot.  Even updating dependencies
risks breaking key functionality in the webapp (see attempts to upgrade
NextJS, React, other dependencies).

This commit freezes dependency management of the web app in order to
focus limited developer resources on replacing the webapp with something
fit for purpose, per the report submitted to MCA w/c 25 Apr.

Co-authored-by: Olly Swanson <42551149+ollyswanson@users.noreply.github.com>