GET Registration by BeaconId and AccountHolderId.
This should not have been implemented in this fashion (i.e. with
accountHolderId as a query parameter), but given the need to fix the
over-fetching of account holder beacons that is done by the webapp, and
the lack of identity in requests sent between the webapp and service,
this was deemed to be the least-bad approach.

Co-authored by: Zack Adlington <54271433+zackads@users.noreply.github.com>